pyproject.toml
[tool.poetry]
name = "thought-CLI"
version = "0.1.0"
description = "A command line tool that can display info from CPX server"
authors = ["cheuk"]

[tool.poetry.dependencies]
python = "^3.7"
click = "^8.1.0"
requests = "^2.24.0"
tabulate = "^0.8.7"
pytest = "^6.1.2"
pytest-mock = "^3.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"

[tool.pytest.ini_options]
pythonpath = [
  ".", "src",
]